+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - module/apmsql: don't report driver.ErrBadConn, context.Canceled (#346, #348)
  - Added ErrorLogRecord.Error field, for associating an error value with a log record (#380)
  - module/apmlogrus: introduce logrus exception-tracking hook, and log correlation (#381)
+ - module/apmbeego: introduce Beego instrumentation module (#386)
 
 ## [v1.0.0](https://github.com/elastic/apm-agent-go/releases/tag/v1.0.0)